+++ pastel/gfx/draw_textured_triangle.hpp
@@ -367,7 +367,7 @@ namespace Pastel
 
 		for (integer y = yMin;y < yMid;++y)
 		{
-			const integer xBegin = std::max(toPixelSpanPoint(xLeft), 0);
+			const integer xBegin = std::max(toPixelSpanPoint(xLeft), (integer)0);
 			const integer xEnd = std::min(toPixelSpanPoint(xRight), width);
 
 			if (xEnd - xBegin > 0)
@@ -449,7 +449,7 @@ namespace Pastel
 
 		for (integer y = yMid;y < yMax;++y)
 		{
-			const integer xBegin = std::max(toPixelSpanPoint(xLeft), 0);
+			const integer xBegin = std::max(toPixelSpanPoint(xLeft), (integer)0);
 			const integer xEnd = std::min(toPixelSpanPoint(xRight), width);
 
 			if (xEnd - xBegin > 0)
